requirements.txt
Flask~=2.3.1
Flask-PyMongo==2.3.0
redis==4.5.4
pymongo~=4.3.3
flask-cors
python-dotenv
python-jose
six